# apifox-openapi.yaml
openapi: 3.0.3
info:
  title: 剧本杀拼车系统 API
  version: 2.4.0
  description: |
    JuPin 后端接口设计文档，可直接导入 Apifox。

    本地测试 Base URL: http://localhost:8080

    前端和 Nginx 未启动不影响接口测试，Apifox 直接请求 Spring Boot 后端即可。

    认证方式：
    - 注册/登录接口不需要 Token
    - 玩家端接口使用玩家 Token
    - 店家端接口使用店家 Token
    - 管理端接口使用管理员 Token
    - Header: Authorization: Bearer <accessToken>

    WebSocket:
    - 连接端点：/ws
    - 拼车状态订阅：/topic/pool/{poolId}
    - 群聊订阅：/topic/pool/{poolId}/chat
    - 匹配推送：/queue/match/{userId}
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
security:
  - bearerAuth: []
paths:
...
  /api/player/pool/create:
    post:
      tags: [玩家端-拼车]
      summary: 发布玩家局
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoolCreateRequest'
            example:
              scriptId: 1
              scriptName: 年轮
              scriptType: 硬核
              roles: '[{"name":"侦探","desc":"推理位"}]'
              city: 上海
              address: 静安区南京西路XX号
              startTime: '2026-06-01 14:00:00'
              endTime: '2026-06-01 18:00:00'
              maxMembers: 6
              price: 88
              deposit: 10
              joinType: 1
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolResult'
  /api/player/pool/list:
    get:
      tags: [玩家端-拼车]
      summary: 玩家端拼车列表
      security: []
      parameters:
        - $ref: '#/components/parameters/City'
        - $ref: '#/components/parameters/ScriptType'
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/PriceMin'
        - $ref: '#/components/parameters/PriceMax'
        - $ref: '#/components/parameters/StartTimeAfter'
        - $ref: '#/components/parameters/StartTimeBefore'
        - $ref: '#/components/parameters/Recommend'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 拼车列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolListResult'
...
  /api/player/shop/list:
    get:
      tags: [玩家端-店铺]
      summary: 玩家端店铺列表
      parameters:
        - $ref: '#/components/parameters/City'
        - name: keyword
          in: query
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 店铺列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopListResult'
  /api/player/shop/{shopId}:
    get:
      tags: [玩家端-店铺]
      summary: 玩家端店铺主页详情
      parameters:
        - $ref: '#/components/parameters/ShopId'
      responses:
        '200':
          description: 店铺详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopResult'
...
  /api/player/pool/{poolId}:
    get:
      tags: [玩家端-拼车]
      summary: 拼车详情
      security: []
      parameters:
        - $ref: '#/components/parameters/PoolId'
      responses:
        '200':
          description: 拼车详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolResult'
  /api/player/pool/{poolId}/price:
    put:
      tags: [玩家端-拼车]
      summary: 修改价格
      parameters:
        - $ref: '#/components/parameters/PoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: number
              format: double
            example: 99
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/player/pool/{poolId}/transfer-dm:
    put:
      tags: [玩家端-拼车]
      summary: 转让 DM
      parameters:
        - $ref: '#/components/parameters/PoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: integer
              format: int64
            example: 2
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/player/pool/{poolId}/join:
    post:
      tags: [玩家端-拼车]
      summary: 申请加入拼车
      parameters:
        - $ref: '#/components/parameters/PoolId'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/player/pool/{poolId}/leave:
    post:
      tags: [玩家端-拼车]
      summary: 退出/跳车
      parameters:
        - $ref: '#/components/parameters/PoolId'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/player/pool/{poolId}/approve/{userId}:
    post:
      tags: [玩家端-拼车]
      summary: 发布人通过申请
      parameters:
        - $ref: '#/components/parameters/PoolId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/player/pool/{poolId}/reject/{userId}:
    post:
      tags: [玩家端-拼车]
      summary: 发布人拒绝申请
      parameters:
        - $ref: '#/components/parameters/PoolId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/player/pool/{poolId}/complete:
    post:
      tags: [玩家端-拼车]
      summary: 发起拼车成功确认
      parameters:
        - $ref: '#/components/parameters/PoolId'
      responses:
        '200':
          description: 确认进度
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmResult'
  /api/player/pool/{poolId}/confirm:
    post:
      tags: [玩家端-拼车]
      summary: 提交确认
      parameters:
        - $ref: '#/components/parameters/PoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmRequest'
            example:
              confirmed: true
      responses:
        '200':
          description: 确认进度
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmResult'
  /api/player/pool/{poolId}/finish:
    post:
      tags: [玩家端-拼车]
      summary: 发起剧本杀完成确认
      parameters:
        - $ref: '#/components/parameters/PoolId'
      responses:
        '200':
          description: 确认进度
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmResult'
  /api/player/pool/{poolId}/members:
    get:
      tags: [玩家端-拼车]
      summary: 拼车成员列表
      parameters:
        - $ref: '#/components/parameters/PoolId'
      responses:
        '200':
          description: 成员列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberListResult'
  /api/player/pool/{poolId}/role/select:
    post:
      tags: [玩家端-拼车]
      summary: 选择剧本角色
      parameters:
        - $ref: '#/components/parameters/PoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleSelectRequest'
            example:
              roleName: 侦探
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/player/pool/{poolId}/roles:
    get:
      tags: [玩家端-拼车]
      summary: 角色列表及选择状态
      parameters:
        - $ref: '#/components/parameters/PoolId'
      responses:
        '200':
          description: 角色选择状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleStatusListResult'
...
  /api/shop/create:
    post:
      tags: [店家端-店铺]
      summary: 创建店铺
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopCreateRequest'
            example:
              name: 静安剧本杀馆
              address: 上海市静安区南京西路XXX号
              phone: 021-12345678
              logo: ''
              cover: ''
              description: 主打硬核推理和情感本
              openingHours: 10:00-22:00
              city: 上海
      responses:
        '200':
          description: 店铺信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopResult'
...
  /api/shop/{shopId}/members:
    get:
      tags: [店家端-店铺]
      summary: 店铺成员列表
      parameters:
        - $ref: '#/components/parameters/ShopId'
      responses:
        '200':
          description: 店铺成员列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopMemberListResult'
...
  /api/shop/pool/create:
    post:
      tags: [店家端-拼车]
      summary: 发布店家局
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoolCreateRequest'
            example:
              shopId: 1
              scriptId: 1
              scriptName: 年轮
              scriptType: 硬核
              city: 上海
              address: 静安区南京西路XX号
              startTime: '2026-06-02 14:00:00'
              endTime: '2026-06-02 18:00:00'
              maxMembers: 6
              price: 88
              deposit: 10
              joinType: 1
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolResult'
  /api/shop/pool/list:
    get:
      tags: [店家端-拼车]
      summary: 店家局列表
      parameters:
        - $ref: '#/components/parameters/City'
        - $ref: '#/components/parameters/ScriptType'
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/PriceMin'
        - $ref: '#/components/parameters/PriceMax'
        - $ref: '#/components/parameters/StartTimeAfter'
        - $ref: '#/components/parameters/StartTimeBefore'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 店家局列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolListResult'
  /api/shop/pool/{poolId}:
    get:
      tags: [店家端-拼车]
      summary: 店家局详情
      parameters:
        - $ref: '#/components/parameters/PoolId'
      responses:
        '200':
          description: 拼车详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolResult'
  /api/shop/pool/{poolId}/assign-dm:
    post:
      tags: [店家端-拼车]
      summary: 指派 DM
      parameters:
        - $ref: '#/components/parameters/PoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignDmRequest'
            example:
              poolId: 1
              dmId: 2
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/shop/pool/{poolId}/complete:
    post:
      tags: [店家端-拼车]
      summary: 发起拼车成功确认
      parameters:
        - $ref: '#/components/parameters/PoolId'
      responses:
        '200':
          description: 确认进度
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmResult'
  /api/shop/pool/{poolId}/confirm:
    post:
      tags: [店家端-拼车]
      summary: 提交确认
      parameters:
        - $ref: '#/components/parameters/PoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmRequest'
            example:
              confirmed: true
      responses:
        '200':
          description: 确认进度
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmResult'
  /api/shop/pool/{poolId}/finish:
    post:
      tags: [店家端-拼车]
      summary: 发起剧本杀完成确认
      parameters:
        - $ref: '#/components/parameters/PoolId'
      responses:
        '200':
          description: 确认进度
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmResult'
  /api/shop/pool/{poolId}/members:
    get:
      tags: [店家端-拼车]
      summary: 店家局成员列表
      parameters:
        - $ref: '#/components/parameters/PoolId'
      responses:
        '200':
          description: 成员列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberListResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PoolId:
      name: poolId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      example: 1
    ShopId:
      name: shopId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      example: 1
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      example: 2
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
      example: 1
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 10
      example: 10
    City:
      name: city
      in: query
      required: false
      schema:
        type: string
      example: 上海
    ScriptType:
      name: scriptType
      in: query
      required: false
      schema:
        type: string
      example: 硬核
    Status:
      name: status
      in: query
      required: false
      schema:
        type: integer
    PriceMin:
      name: priceMin
      in: query
      required: false
      schema:
        type: number
        format: double
    PriceMax:
      name: priceMax
      in: query
      required: false
      schema:
        type: number
        format: double
    StartTimeAfter:
      name: startTimeAfter
      in: query
      required: false
      schema:
        type: string
      example: '2026-06-01 00:00:00'
    StartTimeBefore:
      name: startTimeBefore
      in: query
      required: false
      schema:
        type: string
      example: '2026-06-30 23:59:59'
    Recommend:
      name: recommend
      in: query
      required: false
      schema:
        type: boolean
        default: false
  responses:
    Ok:
      description: 操作成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseResult'
  schemas:
    BaseResult:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          nullable: true
...
    PoolResult:
      allOf:
        - $ref: '#/components/schemas/BaseResult'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PoolVO'
    PoolListResult:
      allOf:
        - $ref: '#/components/schemas/BaseResult'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/PoolVO'
    MemberListResult:
      allOf:
        - $ref: '#/components/schemas/BaseResult'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/MemberVO'
    ConfirmResult:
      allOf:
        - $ref: '#/components/schemas/BaseResult'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ConfirmVO'
...
    ShopResult:
      allOf:
        - $ref: '#/components/schemas/BaseResult'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ShopVO'
    ShopListResult:
      allOf:
        - $ref: '#/components/schemas/BaseResult'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ShopVO'
    ShopMemberListResult:
      allOf:
        - $ref: '#/components/schemas/BaseResult'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ShopMemberVO'
...
    RoleStatusListResult:
      allOf:
        - $ref: '#/components/schemas/BaseResult'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/RoleStatusVO'
...
    PoolCreateRequest:
      type: object
      required: [scriptName, city, startTime, maxMembers]
      properties:
        type:
          type: integer
          description: 0-玩家局 1-店家局
        scriptId:
          type: integer
          format: int64
        shopId:
          type: integer
          format: int64
        scriptName:
          type: string
        scriptType:
          type: string
        roles:
          type: string
          description: 角色列表 JSON 字符串
        city:
          type: string
        address:
          type: string
        startTime:
          type: string
          example: '2026-06-01 14:00:00'
        endTime:
          type: string
          example: '2026-06-01 18:00:00'
        maxMembers:
          type: integer
        price:
          type: number
          format: double
        deposit:
          type: number
          format: double
        joinType:
          type: integer
          description: 0-审核制 1-自动通过
        dmId:
          type: integer
          format: int64
    PoolVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: integer
        ownerId:
          type: integer
          format: int64
        shopId:
          type: integer
          format: int64
        scriptId:
          type: integer
          format: int64
        scriptName:
          type: string
        scriptType:
          type: string
        city:
          type: string
        address:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        maxMembers:
          type: integer
        currentMembers:
          type: integer
        price:
          type: number
        deposit:
          type: number
        dmId:
          type: integer
          format: int64
        joinType:
          type: integer
        status:
          type: integer
          description: 0-开放 1-满员 2-拼车成功 3-剧本杀完成 4-已取消
    MemberVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        nickname:
          type: string
        avatar:
          type: string
        role:
          type: integer
        status:
          type: integer
        selectedRole:
          type: string
        joinTime:
          type: string
    ConfirmRequest:
      type: object
      required: [confirmed]
      properties:
        confirmed:
          type: boolean
    ConfirmVO:
      type: object
      properties:
        poolId:
          type: integer
          format: int64
        confirmedCount:
          type: integer
        totalCount:
          type: integer
        completed:
          type: boolean
    RoleSelectRequest:
      type: object
      required: [roleName]
      properties:
        roleName:
          type: string
    RoleStatusVO:
      type: object
      properties:
        roleName:
          type: string
        selected:
          type: boolean
        userId:
          type: integer
          format: int64
        nickname:
          type: string
...
    ShopCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        logo:
          type: string
        cover:
          type: string
        description:
          type: string
        openingHours:
          type: string
        city:
          type: string
...
    ShopVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        logo:
          type: string
        cover:
          type: string
        description:
          type: string
        openingHours:
          type: string
        city:
          type: string
        status:
          type: integer
    ShopMemberVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        shopId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        nickname:
          type: string
        role:
          type: integer
        status:
          type: integer
...
    AssignDmRequest:
      type: object
      required: [poolId, dmId]
      properties:
        poolId:
          type: integer
          format: int64
        dmId:
          type: integer
          format: int64
...